refactor: move login screen to shared auth
